Added linker flag to the mex complier command for SDMC so that the mex file encodes the relative location of the linux library. The algorithm can now run without setting the LD_LIBRARY_PATH environment variable.
I have only made this change for the case of Linux and Matlab ver
9.10+. I can not test the other two cases so I will leave it up to
others to make similar changes. For the other linux case, I expect
that the same linker flag will work. I am less sure about the
Windows case. If someone updates the other two cases in the future
then it would probably be best to update the error message